docs: card display and logic for every task
