+++ src/taskmodule.js
@@ -1,0 +1,11 @@
+class Task {
+  constructor(title, description, date, priority, completion) {
+    this.title = title;
+    this.description = description;
+    this.date = date;
+    this.priority = priority;
+    this.completion = completion;
+  }
+}
+
+export {Task};
